Supported extended format for specifying member+interface
AllJoyn.js can usually figure out which interface to use based solely on the member name. When this would be ambiguous the member can be specified using an extended format with form { 'member-name':'interface-name' }
The extended format is supported for method calls and signals.
Changed the samples to use the new format.

Change-Id: Iff1d441ebb81d8f5ac63e34941491640890e35b8
Signed-off-by: gburns <gburns@qce.qualcomm.com>