Adding device name filtering to console
The AllJoyn.js console currently connects to the first device it finds
that implements the console interface. This causes problems when there
are multiple devices on the network. This patch fixes the problem by
filtering announcements based on the device name. The device name can
be specified on the command line for the console app and the
alljoyn-js executable on Windows and Linux.

Change-Id: I05a0aca3911a193c45e20829bd19cb642e71c14e
Signed-off-by: gburns <gburns@qce.qualcomm.com>