Add a python console module
Creates an AJS_PyConsole subclass of AJS_Console that makes the
console functionality accessible from python code.

To build the python module, run 'python setup.py build' in the console
directory. The python module is found in
console/build/lib.<platform>/AJSConsole.so

If AJSConsole.so is in the same directory as a python script, it can
be imported with an "import AJSConsole" statement. Then python code
can make calls like "AJSConsole.Connect()" to trigger a connection,
and "AJSConsole.Eval(javascript_string)" to send javascript to
for alljoynjs to evaluate.

This has limited dev testing on 64-bit Linux with Python 2.7

Change-Id: Ieeac7c1317241685302d6c280bb959c2dc72ede3
Signed-off-by: Mathew Martineau <mathewm@qce.qualcomm.com>