ASACORE-1002: Incorrect SendCmd implementation in win32 version
simio.py was updated to support pwm by adding an additional argument. The code changes were made for linux and darwin but not windows.
Also in this change is a normalization of the SendCmd function on all platforms. This function had clearly been written with just one use case (PWM) in mind.

Change-Id: I918e933ca5dc8c71fee15de2a3e39fd43445a4bc
Signed-off-by: gburns <gburns@qce.qualcomm.com>