Arduino Yun lininoIO integration
This is a first pass at integrating AllJoyn.js with Linux user space GPIO for LininoIO on the
Arduino Yun. Current support is limited to read/write of digital GPIO and support for interrupts and PWM.

Debounce is not yet implemented and there a problems with cleanup when a pin is freed.

Plenty of loose ends but good enough for hacks and demos.

Change-Id: I41251e0007ca8be40d769c4c631413f7ad1618e3
Signed-off-by: Greg Burns <gburns@qce.qualcomm.com>