Added support for module loading require()
This change adds support for module loading for the AJ and IO modules. This side-effect is that all
samples now need to specify require("AJ") and/or require("IO") to load the appropriate modules. This
change reduces the memory footprint for applications that don't use AJ or IO. Console support is
always enabled. Also removed a couple of stale code samples.

Change-Id: I6a9c0da38433b62e66041f3d62a8f331f1a0c3ba
Signed-off-by: Greg Burns <gburns@qce.qualcomm.com>