Allows messages to be sent inside CPS callbacks
This change makes a minimal copy of a method call message so that a reply can be constructed. This
allows the method call message to be closed freeing up the I/O buffer to allow messages and signals
to be received while executing CPS callback functions. Currently this only happens when debugging
JavaScript.

Change-Id: Id9e98e4933f5f801688f6fff7e9acaa75cc57364
Signed-off-by: Greg Burns <gburns@qce.qualcomm.com>