AJCORE-1742: Disable JavaScript eval function
To avoid causing unexpected failures in lockdown mode it seems preferable to disabl
eval in all cases. It is generally considered bad practice to call eval. A determined
developer can still evaluate code at runtime using:

f = new Function(eval_this_code);
f()

Where eval_this_code is the JavaScript code to be executed. There is no easy way to
prevent this but this is not a typical programming idiom.

Change-Id: I7328871cc75173c91867a71633f5a61ce44911c4
Signed-off-by: gburns <gburns@qce.qualcomm.com>